implement pairsByQualityAndQuantity function
This implements a new function which not only sorts by Quality, but also
by Quantity in each individual stack.
This is useful for Vendors with limited transactions (i.e. Fences),
and can be used to sell both the highest quality goods (which is already
the case), as well as freeing as many bag slots as possible by
prioritising smaller stacks in the inventory. Dustman already handles
selling parts of larger stacks so "leftover" stacks are not a concern.